pyproject.toml
[project]
name = "biomapper2"
version = "0.1.0"
description = "Unified toolkit for multiomics data harmonization"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "rdkit>=2025.9.1",
    "requests>=2.32.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/biomapper2"]

[tool.black]
line-length = 120

[dependency-groups]
dev = [
    "black>=25.11.0",
    "mypy>=1.18.2",
    "pandas-stubs>=2.3.2.250926",
    "pyright>=1.1.407",
    "pytest>=8.4.2",
    "types-pyyaml>=6.0.12.20250915",
    "types-requests>=2.32.4.20250913",
]
